# ai-meeting-api/src/main/java/com/only/ai/meeting/api/response/MeetingRoomDTO.java
package com.only.ai.meeting.api.response;

import java.util.List;

/**
 * 会议室DTO
 * 
 * @author AI Meeting Team
 * @since 2025-01-27
 */
public class MeetingRoomDTO {
    private Long id;
    private String name;
    private String location;
    private Integer capacity;
    private List<String> equipment;
    private String status;
    private String description;

    // Getters and Setters
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getLocation() {
        return location;
    }

    public void setLocation(String location) {
        this.location = location;
    }

    public Integer getCapacity() {
        return capacity;
    }

    public void setCapacity(Integer capacity) {
        this.capacity = capacity;
    }

    public List<String> getEquipment() {
        return equipment;
    }

    public void setEquipment(List<String> equipment) {
        this.equipment = equipment;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }
}